:truck: start to structure enrichment module
- break huge files into small ones (in order to allow moving these around)
